requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
appdirs
asgi-csrf
asgiref
attrs
backoff
bcrypt
beautifulsoup4
blinker
build
cachetools
certifi
charset-normalizer
chroma-hnswlib
chromadb
click
click-default-group
coloredlogs
dataclasses-json
datasette
Deprecated
distro
fastapi
filelock
Flask
Flask-Cors
flatbuffers
flexcache
flexparser
frozenlist
fsspec
gitdb
GitPython
google-auth
googleapis-common-protos
grpcio
h11
httpcore
httptools
httpx
huggingface-hub
humanfriendly
hupper
idna
importlib_metadata
importlib_resources
itsdangerous
janus
Jinja2
jiter
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kubernetes
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
markdown-it-py
MarkupSafe
marshmallow
mdurl
mergedeep
mmh3
monotonic
mpmath
multidict
mypy-extensions
narwhals
numpy
oauthlib
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
pandas
Pint
platformdirs
pillow
PyPika
pyproject_hooks
python-dateutil
python-dotenv
python-multipart
pytz
PyYAML
regex
requests
requests-oauthlib
rich
rsa
setuptools
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
starlette
streamlit
sympy
tenacity
tiktoken
tokenizers
toml
tqdm
typer
typing-inspect
typing_extensions
tzdata
urllib3
uvicorn
uvloop
watchfiles
websocket-client
websockets
Werkzeug
wheel
wrapt
yarl
gunicorn
flask
flask-cors
gunicorn
langchain
faiss-cpu
openai